🔒️(project) move copilot token request to backend
The Copilot Endpoint URL is not a secret per se.
However, we anticipate that in the future, we will use a Copilot that
requires passing a secret in the Endpoint URL request's header.
Thus, we no longer pass the Copilot Endpoint URL to the client
JavaScript to avoid exposing this secret to Moodle users and make the
request on the backend side instead.